backend/requirements/new.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.19
celery==3.1.17
discogs-client==2.2.0
Django==1.7.7
django-celery==3.1.10
django-cors-headers==1.1.0
django-crontab==0.6.0
django-socketio==0.3.8
djangorestframework==3.1.1
drf-nested-routers==0.9.0
functools32==3.2.3-2
gevent==0.13.6
gevent-socketio==0.2.1
gevent-websocket==0.2.3
greenlet==0.4.5
id3reader==1.53.20070415
kombu==3.0.24
MySQL-python==1.2.5
oauthlib==1.0.1
Pillow==2.9.0
pytz==2015.2
requests==2.7.0
six==1.9.0
sphinx-me==0.3
wheel==0.24.0
yolk==0.4.3
redis==2.10.3